divide PlotActions.py into actions.control, actions.io, ...
