+++ silx/gui/plot/actions/PlotAction.py
@@ -1,0 +1,81 @@
+# coding: utf-8
+# /*##########################################################################
+#
+# Copyright (c) 2004-2017 European Synchrotron Radiation Facility
+#
+# Permission is hereby granted, free of charge, to any person obtaining a copy
+# of this software and associated documentation files (the "Software"), to deal
+# in the Software without restriction, including without limitation the rights
+# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+# copies of the Software, and to permit persons to whom the Software is
+# furnished to do so, subject to the following conditions:
+#
+# The above copyright notice and this permission notice shall be included in
+# all copies or substantial portions of the Software.
+#
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+# THE SOFTWARE.
+#
+# ###########################################################################*/
+"""TODO
+
+The class :class:`PlotAction` help the creation of a qt.QAction associated
+with a :class:`.PlotWidget`.
+
+"""
+
+from __future__ import division
+
+
+__authors__ = ["V.A. Sole", "T. Vincent", "P. Knobel"]
+__license__ = "MIT"
+__date__ = "20/04/2017"
+
+
+from collections import OrderedDict
+import weakref
+from silx.gui import icons
+from silx.gui import qt
+
+
+class PlotAction(qt.QAction):
+    """Base class for QAction that operates on a PlotWidget.
+
+    :param plot: :class:`.PlotWidget` instance on which to operate.
+    :param icon: QIcon or str name of icon to use
+    :param str text: The name of this action to be used for menu label
+    :param str tooltip: The text of the tooltip
+    :param triggered: The callback to connect to the action's triggered
+                      signal or None for no callback.
+    :param bool checkable: True for checkable action, False otherwise (default)
+    :param parent: See :class:`QAction`.
+    """
+
+    def __init__(self, plot, icon, text, tooltip=None,
+                 triggered=None, checkable=False, parent=None):
+        assert plot is not None
+        self._plotRef = weakref.ref(plot)
+
+        if not isinstance(icon, qt.QIcon):
+            # Try with icon as a string and load corresponding icon
+            icon = icons.getQIcon(icon)
+
+        super(PlotAction, self).__init__(icon, text, parent)
+
+        if tooltip is not None:
+            self.setToolTip(tooltip)
+
+        self.setCheckable(checkable)
+
+        if triggered is not None:
+            self.triggered[bool].connect(triggered)
+
+    @property
+    def plot(self):
+        """The :class:`.PlotWidget` this action group is controlling."""
+        return self._plotRef()
